requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
ollama>=0.1.7
mem0ai>=0.1.0
